ABC155D仕掛

+++ Library/lower_bound.cs
@@ -1,9 +1,9 @@
 using System;
-static int lower_bound(long ng,long ok, Func<long, bool> pred)
+static long lower_bound(long ng, long ok, Func<long, bool> pred)
 {
     while (Math.Abs(ng - ok) > 1)
     {
-        int mid = (ok + ng) / 2;
+        long mid = (ok + ng) / 2;
 
         if (pred(mid)) ok = mid;
         else ng = mid;
